src/skillsnap_frontend/src/AI/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
Babel==2.14.0
bce-python-sdk==0.9.7
bcrypt==4.2.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flywheel==0.5.4
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
grpcio==1.73.1
grpcio-status==1.71.2
grpcio-tools==1.71.0
joblib==1.5.1
jsonpath-ng==1.6.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Levenshtein==0.26.1
